Move pointless-las function out of info kernel,
and into kernels/private/PointlessLas.cpp
Addressing review comments of #4178
"Modify pdal info --summary to not fetch the entire file"
