+++ kernels/private/PointlessLas.hpp
@@ -1,5 +1,5 @@
 /******************************************************************************
-* Copyright (c) 2013, Howard Butler (hobu.inc@gmail.com)
+* Copyright (c) 2023, Connor Manning (connor@hobu.co)
 *
 * All rights reserved.
 *
@@ -34,68 +34,17 @@
 
 #pragma once
 
-#include <pdal/Kernel.hpp>
-#include <pdal/PipelineManager.hpp>
-#include <pdal/PointView.hpp>
-#include <pdal/Stage.hpp>
-#include <pdal/util/FileUtils.hpp>
+#include <memory>
+#include <string>
 
-#include "private/PointlessLas.hpp"
-
-#ifdef __clang__
-#pragma GCC diagnostic ignored "-Wtautological-constant-out-of-range-compare"
-#endif
+#include "arbiter/arbiter.hpp"
 
 namespace pdal
 {
 
-class PDAL_DLL InfoKernel : public Kernel
-{
-public:
-    std::string getName() const;
-    int execute(); // overrride
-
-    InfoKernel();
-    void setup(const std::string& filename);
-    MetadataNode run(const std::string& filename);
-
-    inline bool showAll() { return m_showAll; }
-    inline void doShowAll(bool value) { m_showAll = value; }
-    inline void doComputeSummary(bool value) { m_showSummary = value; }
-    inline void doComputeBoundary(bool value) { m_boundary = value; }
-
-private:
-    void addSwitches(ProgramArgs& args);
-    void validateSwitches(ProgramArgs& args);
-    void makeReader(const std::string& filename);
-    void makePipeline();
-    void dump(MetadataNode& root);
-    MetadataNode dumpSummary(const QuickInfo& qi);
-
-    std::string m_inputFile;
-    bool m_showStats;
-    bool m_showSchema;
-    bool m_showAll;
-    bool m_showMetadata;
-    bool m_boundary;
-    bool m_stac;
-    std::string m_pointIndexes;
-    std::string m_dimensions;
-    std::string m_enumerate;
-    std::string m_queryPoint;
-    std::string m_pipelineFile;
-    std::string m_pcType;
-    bool m_showSummary;
-    bool m_needPoints;
-    bool m_usestdin;
-
-    Stage *m_statsStage;
-    Stage *m_hexbinStage;
-    Stage *m_infoStage;
-    Stage *m_reader;
-    Stage *m_stacStage;
-
-    MetadataNode m_tree;
-};
+// Uses arbiter to make range requests to fetch the LAS header and VLRs of a remote LAS file.
+// Returns a LocalHandle to a temporary file that has the same LAS header / VLRs as the remote
+// file (except for changed byte-offsets where necessary) but doesn't contain any points.
+std::unique_ptr<arbiter::LocalHandle> getPointlessLasFile(const std::string& filename);
 
 } // namespace pdal
